Was still failing to start properly at boot; tidied up some superfluous code
